backend/requirements.txt
aiohttp==3.8.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
amqp==5.1.1
appnope==0.1.3
asgiref==3.6.0
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
billiard==3.6.4.0
boto3==1.28.29
botocore==1.31.29
Brotli==1.0.9
CacheControl==0.12.11
cachetools==5.3.0
celery==5.2.7
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
contourpy==1.0.7
coreapi==2.3.3
coreschema==0.0.4
cron-descriptor==1.3.0
cssselect2==0.7.0
cycler==0.11.0
decorator==5.1.1
Django==4.0.10
django-after-response==0.2.2
django-appconf==1.0.5
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-cors-headers==3.14.0
django-db-logger==0.1.12
django-environ==0.8.1
django-filter==21.1
django-imagekit==4.1.0
django-modeltranslation==0.18.9
django-storages==1.13.2
django-timezone-field==5.0
django_debug_toolbar==3.8.1
djangorestframework==3.13.1
djangorestframework-jsonapi==5.0.0
djangorestframework-simplejwt==5.2.2
docopt==0.6.2
drf-yasg==1.21.5
et-xmlfile==1.1.0
executing==1.2.0
fcm-django==1.0.14
firebase-admin==5.4.0
fonttools==4.39.4
freezegun==1.2.2
frozenlist==1.4.0
geographiclib==2.0
geopy==2.3.0
google-api-core==2.11.0
google-api-python-client==2.86.0
google-auth==2.18.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.11.1
google-cloud-storage==2.9.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.0
grpcio==1.54.2
grpcio-status==1.54.2
gunicorn==20.1.0
html5lib==1.1
httplib2==0.22.0
idna==3.4
inflection==0.5.1
ipython==8.13.2
itypes==1.2.0
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
kiwisolver==1.4.4
kombu==5.2.4
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
msgpack==1.0.5
multidict==6.0.4
num2words==0.5.13
numpy==1.24.3
openpyxl==3.1.2
packaging==23.1
pandas==1.5.3
parso==0.8.3
pexpect==4.8.0
phonenumbers==8.13.18
pickleshare==0.7.5
pilkit==2.0
Pillow==9.5.0
prompt-toolkit==3.0.38
proto-plus==1.22.2
protobuf==4.23.0
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydyf==0.6.0
Pygments==2.15.1
PyJWT==2.7.0
pyparsing==3.0.9
pyphen==0.14.0
python-crontab==2.7.1
python-dateutil==2.8.2
pytz==2023.3
redis==4.5.5
requests==2.30.0
rsa==4.9
ruamel.yaml==0.17.26
ruamel.yaml.clib==0.2.7
s3transfer==0.6.2
sentry-sdk==1.28.1
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
tinycss2==1.2.1
traitlets==5.9.0
typing_extensions==4.5.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.15
vine==5.0.0
wcwidth==0.2.6
weasyprint==56.1
webencodings==0.5.1
XlsxWriter==3.1.0
yarl==1.9.2
zopfli==0.2.2
